add notifiction for amin and wh chats

+++ delivery/handlers/admin_handlers.py
@@ -38,28 +38,29 @@ async def close_order(callback: types.CallbackQuery, state: FSMContext, sql_con:
     order_id = re.search("- Трек номер -- .+?;", callback.message.text)[0].split(' -- ')[1][:-1]
     data = sql_con.get_order(order_id)
     sql_con.modify_order(order_id, 'stage', 'closed')
-    new_message = re.sub(".+\n\n", "ЗАКАЗ ЗАКРЫТ\n", callback.message.text)
+    order_message = texts['texts']['order_info'].format(track= data['track_num'], country = data['country'], address = data['address'], password = data['pass'], desc = data['descriprion'], price = data['price'])
+    new_message = f"ЗАКАЗ ЗАКРЫТ\n{order_message}"
 
     if data['delivery_group_messageid'] != None:
         try:
             await bot.delete_message(chat_id = DELIVERY_CHAT, message_id = int(data['delivery_group_messageid']))
             await delete_media(data['delivery_group_mediaid'], DELIVERY_CHAT, bot)
-        except aiogram.exceptions.TelegramBadRequest:
-            pass
+        except aiogram.exceptions.TelegramBadRequest as e:
+            print(e, '--------', order_id)
 
     if data['delivery_private_messageid'] != None:
         if data['deliveryman_id'] != '':
             try:
                 await bot.delete_message(chat_id = data['deliveryman_id'], message_id = int(data['delivery_private_messageid']))
                 await delete_media(data['delivery_private_mediaid'], data['deliveryman_id'], bot)
-            except aiogram.exceptions.TelegramBadRequest:
-                pass
+            except aiogram.exceptions.TelegramBadRequest as e:
+                print(e, '--------', order_id)
     if data['warehouse_messageid'] != None:
         try:
             await bot.delete_message(chat_id = WAREHOUSE_CHATID, message_id = int(data['warehouse_messageid']))
             await delete_media(data['warehouse_mediaid'], WAREHOUSE_CHATID, bot)
-        except aiogram.exceptions.TelegramBadRequest:
-                pass
+        except aiogram.exceptions.TelegramBadRequest as e:
+            print(e, '--------', order_id)
 
     await callback.message.edit_text(text=new_message)
     await callback.answer()
@@ -71,29 +72,34 @@ async def order_not_delivered(callback: types.CallbackQuery, state: FSMContext, 
     order_id = re.search("- Трек номер -- .+?;", callback.message.text)[0].split(' -- ')[1][:-1]
     data = sql_con.get_order(order_id)
     sql_con.modify_order(order_id, 'stage', 'closed')
-    new_message_to_all = re.sub(".+\n\n", "ЗАКАЗ ЗАКРЫТ\n", callback.message.text)
-    new_message_to_admin = re.sub(".+\n\n", "#not_delivered\nПосылка не поступила\n", callback.message.text)
+    
+    order_message = texts['texts']['order_info'].format(track= data['track_num'], country = data['country'], address = data['address'], password = data['pass'], desc = data['descriprion'], price = data['price'])
+    new_message_to_all = f"ЗАКАЗ ЗАКРЫТ\n{order_message}"
+    new_message_to_admin = f"#not_delivered\nПосылка не поступила\n{order_message}"
 
-    if data['delivery_group_messageid'] != None:
-        try:
-            await bot.delete_message(chat_id = DELIVERY_CHAT, message_id = int(data['delivery_group_messageid']))
-            await delete_media(data['delivery_group_mediaid'], DELIVERY_CHAT, bot)
-        except aiogram.exceptions.TelegramBadRequest:
-            pass
-
-    if data['delivery_private_messageid'] != None:
-        if data['deliveryman_id'] != '':
+    if data != None:
+        if data['delivery_group_messageid'] != None:
             try:
-                await bot.delete_message(chat_id = data['deliveryman_id'], message_id = int(data['delivery_private_messageid']))
-                await delete_media(data['delivery_private_mediaid'], data['deliveryman_id'], bot)
-            except aiogram.exceptions.TelegramBadRequest:
-                pass
-    if data['warehouse_messageid'] != None:
-        try:
-            await bot.delete_message(chat_id = WAREHOUSE_CHATID, message_id = int(data['warehouse_messageid']))
-            await delete_media(data['warehouse_mediaid'], WAREHOUSE_CHATID, bot)
-        except aiogram.exceptions.TelegramBadRequest:
-                pass
+                await bot.delete_message(chat_id = DELIVERY_CHAT, message_id = int(data['delivery_group_messageid']))
+                await delete_media(data['delivery_group_mediaid'], DELIVERY_CHAT, bot)
+            except aiogram.exceptions.TelegramBadRequest as e:
+                print(e, '--------', order_id)
+
+        if data['delivery_private_messageid'] != None:
+            if data['deliveryman_id'] != '':
+                try:
+                    await bot.delete_message(chat_id = data['deliveryman_id'], message_id = int(data['delivery_private_messageid']))
+                    await delete_media(data['delivery_private_mediaid'], data['deliveryman_id'], bot)
+                except aiogram.exceptions.TelegramBadRequest as e:
+                    print(e, '--------', order_id)
+        if data['warehouse_messageid'] != None:
+            try:
+                await bot.delete_message(chat_id = WAREHOUSE_CHATID, message_id = int(data['warehouse_messageid']))
+                await delete_media(data['warehouse_mediaid'], WAREHOUSE_CHATID, bot)
+            except aiogram.exceptions.TelegramBadRequest as e:
+                print(e, '--------', order_id)
+    else:
+        print(f'NOT DATA FOR ORDER ----- {order_id}')
 
     await callback.message.edit_text(text=new_message_to_admin)
     await callback.answer()
